# docs/swagger/swagger.yaml
openapi: 3.0.0
info:
  title: Book Sharing Application
  version: 1.0.0
  contact:
    name: Pritam Singh
    url: https://github.com/Zzocker
    email: pkspritam10@gmail.com
servers:
  - url: http://localhost:9001/api/v1
...
security:
  - authToken: []
components:
  securitySchemes:
    authToken:
      type: http
      scheme: bearer
  schemas:
...
    registerUserReq:
      type: object
      required:
        - username
        - email
        - name
        - password
      properties:
        username:
          type: string
          example: Zzocker
        email:
          type: string
          example: something@gmail.com
        name:
          type: string
          example: First Last Name
        password:
          type: string
          example: pw
        profile_pic:
          type: string
          format: binary
    userProfile:
      type: object
      properties:
        username:
          type: string
          example: Zzocker
        email:
          type: string
          example: something@gmail.com
        name:
          type: string
          example: First Last Name
        profile_pic:
          type: string
          format: binary
    updateUserProfileReq:
      type: object
      properties:
        username:
          type: string
        email:
          type: string
        name:
          type: string
        password:
          type: string
        profile_pic:
          type: string
          format: binary
    profilePic:
      type: object
      properties:
        data:
          type: string
          format: base64
          description: base64 encoded image file
    ### books
...
paths:
...
  /user_profile:
    post:
      tags:
        - User-Profile
      security: []
      description: "Register a new user"
      summary: Register a new user
      requestBody:
        content:
          multipart/form-data:
            schema:
              $ref: "#/components/schemas/registerUserReq"

      responses:
        201:
          description: Success
    get:
      tags:
        - User-Profile
      description: Query users
      summary: Query users
      parameters:
        - name: name
          in: query
          required: false
          schema:
            type: string
            example: First Name
        - name: email
          in: query
          required: false
          schema:
            type: string
            example: something@gmail.com
      responses:
        200:
          description: Success
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/userProfile"

    patch:
      tags:
        - User-Profile
      description: Update User's profile
      summary: Update User's profile
      requestBody:
        content:
          multipart/form-data:
            schema:
              $ref: "#/components/schemas/updateUserProfileReq"

      responses:
        201:
          description: Success
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/userProfile"
    delete:
      tags:
        - User-Profile
      description: Deactivate user profile
      summary: Deactivate user profile
      responses:
        204:
          description: Success
  /user_profile/{username}:
    get:
      tags:
        - User-Profile
      description: Get a user's details
      summary: Get a user's details
      parameters:
        - name: username
          in: path
          required: true
          schema:
            type: string
            example: Zzocker
      responses:
        200:
          description: Success
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/userProfile"
  /user_profile/pic/{username}:
    get:
      tags:
        - User-Profile
      description: "Get profile picture of an user"
      summary: Get profile picture of an user
      parameters:
        - name: username
          in: path
          required: true
          schema:
            type: string
            example: Zzocker
      responses:
        200:
          description: SUCCESS
          content:
            application/json: 
              schema:
                $ref : "#/components/schemas/profilePic"
...